Update patch set 1
Patch Set 1:

Removing this call is debatable, depending on the direction of m7-common/overlays that the maintainers want to take. Either have the overlays device specific or common, and the placement of the overlays, either device specific or common. Been trying to get in touch with a m7-common maintainer on irc the past couple of days, but no one answers. I blame the holidays.

Patch-set: 1